Refactor account and settings screens
Replaced LogoutScreen with a new AccountScreen for user profile management and moved account-related logic from SettingsScreen to AccountScreen. Updated main.dart and burger_menu.dart to route '/account' to AccountScreen. Simplified SettingsScreen to a placeholder and improved planner term screen header layout.
