+++ lib/main.dart
@@ -29,7 +29,7 @@ import 'viewmodels/holidays/holidays_viewmodel.dart';
 import 'views/auth/login_screen.dart';
 import 'views/auth/biometric_screen.dart';
 import 'views/auth/signup_screen.dart';
-import 'views/auth/logout_screen.dart';
+import 'views/auth/account_screen.dart';
 import 'views/planner/planner_screen.dart';
 import 'views/today/today_screen.dart';
 import 'views/holidays/holidays_screen.dart';
@@ -213,9 +213,7 @@ class AceUpApp extends StatelessWidget {
     '/assignments': (context) => const AssignmentsScreen(),
     '/settings': (context) => const SettingsScreen(),
     '/group-stats': (context) => const GroupStatsScreen(),
-    '/account': (context) => LogoutScreen(
-          vm: VmScope.of(context).get<LoginViewModel>(),
-        ),
+    '/account': (context) => AccountScreen(),
     '/signup': (context) {
       return ChangeNotifierProvider(
         create: (_) => SignUpViewModel(),
